Opens a dialog for choosing which numberset to edit
Selecting a set has no effect yet, create set page is always opened with
empty set.

Also fixed the position of edit button and theme switch.

Also changed the text in the dialog to slect the number set to use so it
can be easily be told apart from the edit dialog.

Also added qsTr to button texts on settings page.