Restored the selection dialogs on the settings page
..that stopped working *on Harmattan* after switching to com.nokia.meego 1.0

As this is a simpler approach that does the job, (and for keeping the version
differences to a minimum for easier maintanance), it is used also in Fremantle.

Replaced the Numbersetnamesmodels with a QML ListModel, a funtion in
SettingsHandler that fetches the name list, and a javascript function
that copies the list into the ListModel. All SelectionDialogs now use
the same model, which is populated with an up to date list (with default
included if needed) before each use.

Conflicts:
	CountYourScores.pro
	numbersetnamesmodel.h
	qml/CountYourScores/SettingsPage.qml