Corrections on the about page
Removed the double occurrence of the line with the link to the source
code.

Chnaged the link to the source code to point directly to the fremantle
branch, where the source code of the Maemo 5 version is located.

Updated the address of instructions to where github now redirects the
old one. (Related to changes on GitHub servers).