Creating new sets with empty name prevented
Creating new sets with an empty string as the name is prevented. Just to
be on the safe side, as keys should never be empty in QSettings. The
problem with deleting such sets was solved earlier, this is to avoid
possible other problems...

Editing an old set with an empty string as the name is still allowed,
for user convenience, and the set name needs not to be changed for
saving. Saving with other name does not remove the original, so forcing
to change the name would not get rid of potential problems anyway.