Added menu on mainpage
Added PullDownMenu with About, Settings, Clear all, Add game and Undo
functions on main page.

For it to work some space between "Player 1 ... " row and top of page
was needed. Propably because the menu is accessed on emulator by pulling
way below the menu glow and a MouseArea there interferes with
(overrides) this. Hopefully this is only a problem in the emulator, not
on the device. (On emulator there is no  bezel, so the gesture starting
from upper bezel must start lower, likely forcing the "grapping area" of
PullDownMenu lower.)